+++ backend/borg/mockborg/mockborg.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=backend/borg/borg.go -destination=backend/borg/mock/mock_borg.go
+//	mockgen -source=backend/borg/borg.go -destination=backend/borg/mockborg/mockborg.go --package=mockborg
 //
 
-// Package mock_borg is a generated GoMock package.
-package mock_borg
+// Package mockborg is a generated GoMock package.
+package mockborg
 
 import (
 	borg "arco/backend/borg"
